Unnecessary cast

+++ src/vector/blas/dot.cpp
@@ -16,12 +16,12 @@ namespace dd_avx{
 		{
 			int thN = omp_get_thread_num();
 			int i=0, is=0, ie=0;
-			get_isie((int)y.size(), is, ie);
+			get_isie(y.size(), is, ie);
 
 			reg r_hi = regs.zeros;
 			reg r_lo = regs.zeros;
 
-			for(i = is; (int)i < (int)(ie-SIMD_Length+1); i += SIMD_Length){
+			for(i = is; i < (ie-SIMD_Length+1); i += SIMD_Length){
 
 				reg x_hi = load(x.hi[i]);
 				reg x_lo = load(x.lo[i]);
@@ -66,7 +66,7 @@ namespace dd_avx{
 			reg r_hi = regs.zeros;
 			reg r_lo = regs.zeros;
 
-			for(i = is; (int)i < (int)(ie-SIMD_Length+1); i += SIMD_Length){
+			for(i = is; i < (ie-SIMD_Length+1); i += SIMD_Length){
 
 				reg x_hi = load(x.data()[i]);
 
@@ -109,7 +109,7 @@ namespace dd_avx{
 			reg r_hi = regs.zeros;
 			reg r_lo = regs.zeros;
 
-			for(i = is; (int)i < (int)(ie-SIMD_Length+1); i += SIMD_Length){
+			for(i = is; i < (ie-SIMD_Length+1); i += SIMD_Length){
 
 				reg x_hi = load(x.hi[i]);
 				reg x_lo = load(x.lo[i]);
@@ -153,7 +153,7 @@ namespace dd_avx{
 			reg r_hi = regs.zeros;
 			reg r_lo = regs.zeros;
 
-			for(i = is; (int)i < (int)(ie-SIMD_Length+1); i += SIMD_Length){
+			for(i = is; i < (ie-SIMD_Length+1); i += SIMD_Length){
 
 				reg x_hi = load(x.data()[i]);
 				reg x_lo = regs.zeros;
